requirements.txt
DrissionPage==4.1.0.9
psutil==6.1.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9